ai-compliance-demo/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.4
Faker==25.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0
